WIP update - In UIContainer::SetDefinitions(), added try/catch block to script   loading - Added SOL_ALL_SAFETIES_ON macro to cmake scripts - Reordered the sol::state and UIContainer objects of Mapper class so   that the host table of UIContainer is destroyed before the sol::state - Added missing requires clauses to RC class. passing a destroy function   to shared_res of non-pointer resource type MUST take a reference   parameter. Passing by value, even if otherwise valid, will result in   compile error.

+++ ProtoMapper/include/ui/Font.hpp
@@ -35,7 +35,7 @@
 
 namespace proto
 {
-    enum class FontStyle : uint32_t
+    enum class FontStyle : uint8_t
     {
         None = 0u,
         Normal = 1u,
